pyproject.toml
[project]
name = "pa-sample-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "langchain>=1.0.5",
    "langchain-anthropic>=1.0.3",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "mcp>=1.21.0",
    "pattern-agentic-settings>=1.0.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
    "pattern-agentic-messaging @ git+ssh://git@github.com/pattern-agentic/lib-python-messaging.git"
 #   "slim-bindings>=0.6.3,<0.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pa_sample_agent"]
